backend/requirements.txt
django>=3.2
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
python-dotenv
pillow
groq
